refactor(ui): move checkDependencies to ui, accept framework-specific checks via prop
